move paused query and handle logic to workload group manager
