+++ be/src/pipeline/task_scheduler.cpp
@@ -199,169 +199,6 @@ void TaskScheduler::_do_work(size_t index) {
     }
 }
 
-void TaskScheduler::add_paused_task(PipelineTask* task) {
-    std::lock_guard<std::mutex> lock(_paused_queries_lock);
-    auto query_ctx_sptr = task->runtime_state()->get_query_ctx()->shared_from_this();
-    DCHECK(query_ctx_sptr != nullptr);
-    auto wg = query_ctx_sptr->workload_group();
-    auto&& [it, inserted] = _paused_queries_list[wg].emplace(std::move(query_ctx_sptr));
-    if (inserted) {
-        LOG(INFO) << "here insert one new paused query: " << it->query_id()
-                  << ", wg: " << (void*)(wg.get());
-    }
-
-    _paused_queries_cv.notify_all();
-}
-
-/**
- * Strategy 1: A revocable query should not have any running task(PipelineTask).
- * strategy 2: If the workload group is below low water mark, we make all queries in this wg runnable.
- * strategy 3: Pick the query which has the max revocable size to revoke memory.
- * strategy 4: If all queries are not revocable and they all have not any running task,
- *             we choose the max memory usage query to cancel.
- */
-void TaskScheduler::_paused_queries_handler() {
-    while (!_need_to_stop) {
-        {
-            std::unique_lock<std::mutex> lock(_paused_queries_lock);
-            if (_paused_queries_list.empty()) {
-                _paused_queries_cv.wait(lock, [&] { return !_paused_queries_list.empty(); });
-            }
-
-            if (_need_to_stop) {
-                break;
-            }
-
-            if (_paused_queries_list.empty()) {
-                continue;
-            }
-
-            for (auto it = _paused_queries_list.begin(); it != _paused_queries_list.end();) {
-                auto& queries_list = it->second;
-                const auto& wg = it->first;
-                if (queries_list.empty()) {
-                    LOG(INFO) << "wg: " << wg->debug_string() << " has no paused query";
-                    it = _paused_queries_list.erase(it);
-                    continue;
-                }
-
-                bool is_low_wartermark = false;
-                bool is_high_wartermark = false;
-
-                wg->check_mem_used(&is_low_wartermark, &is_high_wartermark);
-
-                if (!is_low_wartermark && !is_high_wartermark) {
-                    LOG(INFO) << "**** there are " << queries_list.size() << " to resume";
-                    for (const auto& query : queries_list) {
-                        LOG(INFO) << "**** resume paused query: " << query.query_id();
-                        query.query_ctx->set_memory_sufficient(true);
-                    }
-
-                    queries_list.clear();
-                    it = _paused_queries_list.erase(it);
-                    continue;
-                } else {
-                    ++it;
-                }
-
-                std::shared_ptr<QueryContext> max_revocable_query;
-                std::shared_ptr<QueryContext> max_memory_usage_query;
-                std::shared_ptr<QueryContext> running_query;
-                bool has_running_query = false;
-                size_t max_revocable_size = 0;
-                size_t max_memory_usage = 0;
-                auto it_to_remove = queries_list.end();
-
-                for (auto query_it = queries_list.begin(); query_it != queries_list.end();) {
-                    const auto& query_ctx = query_it->query_ctx;
-                    size_t revocable_size = 0;
-                    size_t memory_usage = 0;
-                    bool has_running_task = false;
-
-                    if (query_ctx->is_cancelled()) {
-                        LOG(INFO) << "query: " << print_id(query_ctx->query_id())
-                                  << "was canceled, remove from paused list";
-                        query_it = queries_list.erase(query_it);
-                        continue;
-                    }
-
-                    query_ctx->get_revocable_info(&revocable_size, &memory_usage,
-                                                  &has_running_task);
-                    if (has_running_task) {
-                        has_running_query = true;
-                        running_query = query_ctx;
-                        break;
-                    } else if (revocable_size > max_revocable_size) {
-                        max_revocable_query = query_ctx;
-                        max_revocable_size = revocable_size;
-                        it_to_remove = query_it;
-                    } else if (memory_usage > max_memory_usage) {
-                        max_memory_usage_query = query_ctx;
-                        max_memory_usage = memory_usage;
-                        it_to_remove = query_it;
-                    }
-
-                    ++query_it;
-                }
-
-                if (has_running_query) {
-                    LOG(INFO) << "has running task, query: " << print_id(running_query->query_id());
-                    std::this_thread::sleep_for(std::chrono::milliseconds(500));
-                } else if (max_revocable_query) {
-                    queries_list.erase(it_to_remove);
-                    queries_list.insert(queries_list.begin(), max_revocable_query);
-
-                    auto revocable_tasks = max_revocable_query->get_revocable_tasks();
-                    DCHECK(!revocable_tasks.empty());
-
-                    LOG(INFO) << "query: " << print_id(max_revocable_query->query_id()) << ", has "
-                              << revocable_tasks.size()
-                              << " tasks to revoke memory, max revocable size: "
-                              << max_revocable_size;
-                    SCOPED_ATTACH_TASK(max_revocable_query.get());
-                    for (auto* task : revocable_tasks) {
-                        auto st = task->revoke_memory();
-                        if (!st.ok()) {
-                            max_revocable_query->cancel(st);
-                            break;
-                        }
-                    }
-                } else if (max_memory_usage_query) {
-                    bool new_is_low_wartermark = false;
-                    bool new_is_high_wartermark = false;
-                    const auto query_id = print_id(max_memory_usage_query->query_id());
-                    wg->check_mem_used(&new_is_low_wartermark, &new_is_high_wartermark);
-                    if (new_is_high_wartermark) {
-                        if (it_to_remove->elapsed_time() < 2000) {
-                            LOG(INFO) << "memory insufficient and cannot find revocable query, "
-                                         "the max usage query: "
-                                      << query_id << ", usage: " << max_memory_usage
-                                      << ", elapsed: " << it_to_remove->elapsed_time()
-                                      << ", wg info: " << wg->debug_string();
-                            continue;
-                        }
-                        max_memory_usage_query->cancel(Status::InternalError(
-                                "memory insufficient and cannot find revocable query, cancel "
-                                "the "
-                                "biggest usage({}) query({})",
-                                max_memory_usage, query_id));
-                        queries_list.erase(it_to_remove);
-
-                    } else {
-                        LOG(INFO) << "non high water mark, resume "
-                                     "the query: "
-                                  << query_id << ", usage: " << max_memory_usage
-                                  << ", wg info: " << wg->debug_string();
-                        max_memory_usage_query->set_memory_sufficient(true);
-                        queries_list.erase(it_to_remove);
-                    }
-                }
-            }
-        }
-        std::this_thread::sleep_for(std::chrono::milliseconds(100));
-    }
-}
-
 void TaskScheduler::stop() {
     if (!_shutdown) {
         if (_task_queue) {
@@ -369,7 +206,6 @@ void TaskScheduler::stop() {
         }
         if (_fix_thread_pool) {
             _need_to_stop = true;
-            _paused_queries_cv.notify_all();
             _fix_thread_pool->shutdown();
             _fix_thread_pool->wait();
         }
